+++ pages/views/ingredients/mgmGeneral.php
@@ -14,6 +14,17 @@ $res_ingCategory = mysqli_query($conn, "SELECT id,image,name,notes FROM ingCateg
 $res_ingProfiles = mysqli_query($conn, "SELECT id,name FROM ingProfiles ORDER BY id ASC"); //PUBLIC
 $ing = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM ingredients WHERE id = '$ingID' AND owner_id = '$userID'"));
 
+// Generate array with labels data 
+$labelsData = array(); 
+$labelsQ = mysqli_query($conn,"SELECT label_name FROM ingredientLabels WHERE ingredient_id = '$ingID' AND owner_id = '$userID'");
+while($qlabels = mysqli_fetch_array($labelsQ)){
+	
+	$labels = $qlabels['label_name'];
+	array_push($labelsData, $labels); 
+	error_log("Label: " . $labels);
+}
+
+
 if($_GET["newIngName"]){
 	$newIngName = mysqli_real_escape_string($conn, base64_decode($_GET["newIngName"]));
 
@@ -125,8 +136,7 @@ if($_GET["newIngName"]){
     </select>  
   </div>
   <div class="form-floating mt-3 col-md-6">
-    <input name="odor" id="odor" type="text" class="form-control" placeholder="Odor" value="<?php echo $ing['odor']; ?>"/>
-    <label class="mx-2" for="odor">Odor</label>
+	<input type="text" class="form-control" id="labelsinput" placeholder="Start typing to add a label..." data-role="labelsinput" value="<?= implode(",", $labelsData) ?>">
   </div>
 
   <div class="form-floating mt-3 col-12">
@@ -175,7 +185,6 @@ $(document).ready(function() {
 				strength: $("#strength").val(),
 				category: $("#category").val(),
 				physical_state: $("#physical_state").val(),
-				odor: $("#odor").val(),
 				notes: $("#notes").val(),
 				<?php if($ing['name']){?>
 					ing: '<?=$ing['name'];?>'
@@ -223,5 +232,60 @@ $(document).ready(function() {
 		}
 		$('.selectpicker').selectpicker('refresh');
 	});
+
+	$('#labelsinput').on('beforeItemAdd', function(event) {
+	   var label = event.item;   
+	   $.ajax({ 
+			url: '/core/core.php', 
+			type: 'POST',
+			data: {
+				action: "inglabeladd",
+				id: myIngID,
+				label: label
+			},
+			dataType: 'json',
+			success: function (data) {
+				if(data.error){
+					$('#labelsinput').tagsinput('remove', label, {preventPost: true});
+					//$('#set_msg').html(data.error);
+				}
+			},
+			error: function (xhr, status, error) {
+				$('#toast-title').html('<i class="fa-solid fa-circle-exclamation mx-2"></i>An ' + status + ' occurred, check server logs for more info. '+ error);
+				$('.toast-header').removeClass().addClass('toast-header alert-danger');
+				$('.toast').toast('show');
+			}
+		});
+	});
+	
+	$('#labelsinput').val('<?= implode(",", $labelsData) ?>');
+	$('#labelsinput').tagsinput('refresh');
+
+	$('#labelsinput').on('beforeItemRemove', function(event) {
+	var label = event.item;
+
+	$.ajax({ 
+			url: '/core/core.php', 
+			type: 'POST',
+			data: {
+				action: "inglabelremove",
+				id: myIngID,
+				label: label
+			},
+			dataType: 'json',
+			success: function (data) {
+				if (data.error) {
+					$('#labelsinput').tagsinput('add', label, { preventPost: true });
+					$('#set_msg').html('<div class="alert alert-danger mx-2"><i class="fa-solid fa-circle-exclamation mx-2"></i> ' + data.error + '</div>');
+				}
+			},
+			error: function (xhr, status, error) {
+				$('#toast-title').html('<i class="fa-solid fa-circle-exclamation mx-2"></i>An ' + status + ' occurred, check server logs for more info. '+ error);
+				$('.toast-header').removeClass().addClass('toast-header alert-danger');
+				$('.toast').toast('show');
+			}
+		});
+	});
+
 });//end doc
 </script>